Add docstrings/comments, move to utils file
